tempmon/requirements.txt
bs4
flask
python-nmap